extract the signed key in email creation
this is to remove duplicate code with the unit tests, but also makes
more sense: we do not expose a signature that could be potentially
non-verified against the email of the user.

however, it means we do not print the signed key in the clear anymore.
instead, we show the email message, which also makes sense: the user
may want to send mail through another channel.

if the user wants to have a signature exported instead, he/she can use
the -l option to have the signature imported in the local keyring, in
which case it is non-exportable, which also makes sense.