try to rename to monkeysign/monkeyscan
the idea here is that people are confused that when they install
monkeysign, there's no monkeysign binary.

furthermore, the core of the functionality currently resides in
monkeysign, while monkeyscan is still radically different: it doesn't
handle arbitrary key ids (ie. it can't sign a key id you provide, only
one found in a qrcode) and is less tested

ideally both commands would be merged and there would be a -c argument
to monkeysign that would make it stay on the commandline, but as long
as I cannot easily sign arbitrary keys with monkeyscan, i think they
should remain separate.