use exceptions to carry some error conditions in the gpg library
otherwise we weren't catching the error there, as we assume this
returns a string. furthermore, we could have checked the return value,
but we couldn't provide more information to the user as to the details
of the error condition without going into the internals of the gpg
library, something we want to avoid.

exceptions allow us to propagate error messages up in the stack more
easily.