gpg: raise an exception in expect(), as documented, if string not found
this makes more sense: if we expect something, and gpg gives us
something else, we should bail violently and verbosely, as it makes
debugging much easier.

we handle the only case this was actually happening in the unit test
suite properly too.

this was useful to implement deluid support, upcoming.