properly set GPG_TTY if it is missing (Closes: #719908)
This is necessary because gpg seemingly determine the terminal when
called from monkeysign. The reasons behind this are obscure and I
couldn't understand why, but the above fix works, at least.

In the gpg-agent manual, users are explicitely told to configure the
GPG_TTY variable in their session yet I somewhat missed this myself
and I suspect a broad range of users are not doing that, gpg-agent now
being automatically configured by graphical session managers and so
on.

So let's help our poor users a little here.