gpg: Implemented revoked for OpenPGP Keys
The property itself is straight forward to implement, because we already
have parsed the information from the --list-keys call.
However, GnuPG does not export that information for secret keys.  Hence,
if you wanted to know whether the secret key at your hand is revoked,
you need to find the public key first.