gpg: Implemented revoked for OpenPGP UIDs
We already have the information whether the UID has been revoked.  It
came as the trust attribute and returning that is straight forward.
However, for secret keys, the information is not exported and you need
to find the appropriate public key in order to determine whether the UID
is revoked.